% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoOL_functions.R
\name{CoOL_0_binary_encode_exposure_data}
\alias{CoOL_0_binary_encode_exposure_data}
\title{Binary encode exposure data}
\usage{
CoOL_0_binary_encode_exposure_data(exposure_data)
}
\arguments{
\item{exposure_data}{The exposure data set.}
}
\value{
Data frame with the expanded exposure data, where all variables are binary encoded.
}
\description{
This function binary encodes the exposure data set so that each category is coded 0 and 1 (e.g. the variable sex will be two variables men (1/0) and women (0/1)).
}
\examples{
#See the example under CoOL_0_working_example

}
\references{
Rieckmann, Dworzynski, Arras, Lapuschkin, Samek, Arah, Rod, Ekstrom. 2022. Causes of outcome learning: A causal inference-inspired machine learning approach to disentangling common combinations of potential causes of a health outcome. International Journal of Epidemiology <https://doi.org/10.1093/ije/dyac078>
}
