% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoOL_functions.R
\name{CoOL_0_common_simulation}
\alias{CoOL_0_common_simulation}
\title{Common example}
\usage{
CoOL_0_common_simulation(n)
}
\arguments{
\item{n}{number of observations for the synthetic data}
}
\description{
To reproduce the common causes example.
}
