% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoOL_functions.R
\name{CoOL_0_complex_simulation}
\alias{CoOL_0_complex_simulation}
\title{Complex example}
\usage{
CoOL_0_complex_simulation(n)
}
\arguments{
\item{n}{number of observations for the synthetic data}
}
\description{
To reproduce the complex example.
}
