% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoOL_functions.R
\name{CoOL_0_confounding_simulation}
\alias{CoOL_0_confounding_simulation}
\title{Confounding example}
\usage{
CoOL_0_confounding_simulation(n)
}
\arguments{
\item{n}{number of observations for the synthetic data}
}
\description{
To reproduce the confounding example.
}
