% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoOL_functions.R
\name{CoOL_0_m_bias_simulation}
\alias{CoOL_0_m_bias_simulation}
\title{M-bias example}
\usage{
CoOL_0_m_bias_simulation(n)
}
\arguments{
\item{n}{number of observations for the synthetic data}
}
\description{
To reproduce the M-bias example.
}
