% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoOL_functions.R
\name{CoOL_0_mediation_simulation}
\alias{CoOL_0_mediation_simulation}
\title{Mediation example}
\usage{
CoOL_0_mediation_simulation(n)
}
\arguments{
\item{n}{number of observations for the synthetic data}
}
\description{
To reproduce the mediation example.
}
\examples{


}
