% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoOL_functions.R
\name{CoOL_4_AUC}
\alias{CoOL_4_AUC}
\title{Plot the ROC AUC}
\usage{
CoOL_4_AUC(
  outcome_data,
  exposure_data,
  model,
  title = "Receiver operating\\ncharacteristic curve",
  restore_par_options = TRUE
)
}
\arguments{
\item{outcome_data}{The outcome data.}

\item{exposure_data}{The exposure data.}

\item{model}{The fitted the non-negative neural network.}

\item{title}{Title on the plot.}

\item{restore_par_options}{Restore par options.}
}
\value{
A plot of the ROC and the ROC AUC value.
}
\description{
Plot the ROC AUC
}
\examples{
#See the example under CoOL_0_working_example
}
\references{
Rieckmann, Dworzynski, Arras, Lapuschkin, Samek, Arah, Rod, Ekstrom. 2022. Causes of outcome learning: A causal inference-inspired machine learning approach to disentangling common combinations of potential causes of a health outcome. International Journal of Epidemiology <https://doi.org/10.1093/ije/dyac078>
}
