% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoOL_functions.R
\name{CoOL_6_calibration_plot}
\alias{CoOL_6_calibration_plot}
\title{Calibration curve}
\usage{
CoOL_6_calibration_plot(
  exposure_data,
  outcome_data,
  model,
  sub_groups,
  ipw = 1,
  restore_par_options = TRUE
)
}
\arguments{
\item{exposure_data}{The exposure dataset.}

\item{outcome_data}{The outcome vector.}

\item{model}{The fitted non-negative neural network.}

\item{sub_groups}{The vector with the assigned sub_group numbers.}

\item{ipw}{a vector of weights per observation to allow for inverse probability of censoring weighting to correct for selection bias}

\item{restore_par_options}{Restore par options.}
}
\value{
A calibration curve.
}
\description{
Shows the calibration curve e.i. the predicted risk vs the actual risk by subgroups.
}
\examples{
#See the example under CoOL_0_working_example
}
\references{
Rieckmann, Dworzynski, Arras, Lapuschkin, Samek, Arah, Rod, Ekstrom. 2022. Causes of outcome learning: A causal inference-inspired machine learning approach to disentangling common combinations of potential causes of a health outcome. International Journal of Epidemiology <https://doi.org/10.1093/ije/dyac078>
}
