% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoOL_functions.R
\name{CoOL_6_individual_effects_matrix}
\alias{CoOL_6_individual_effects_matrix}
\title{Risk contribution matrix based on individual effects (had all other exposures been set to zero)}
\usage{
CoOL_6_individual_effects_matrix(X, model)
}
\arguments{
\item{X}{The exposure data.}

\item{model}{The fitted the non-negative neural network.}
}
\description{
Estimating the risk contribution for each exposure if each individual had been exposed to only one exposure, with the value the individual actually had.
}
\examples{
#See the example under CoOL_0_working_example
}
\references{
Rieckmann, Dworzynski, Arras, Lapuschkin, Samek, Arah, Rod, Ekstrom. Causes of outcome learning: A causal inference-inspired machine learning approach to disentangling common combinations of potential causes of a health outcome. medRxiv (2020) <doi:10.1101/2020.12.10.20225243>
}
