% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoOL_functions.R
\name{CoOL_6_number_of_sub_groups}
\alias{CoOL_6_number_of_sub_groups}
\title{Number of subgroups}
\usage{
CoOL_6_number_of_sub_groups(
  risk_contributions,
  low_number = 1,
  high_number = 5,
  ipw = 1,
  restore_par_options = TRUE
)
}
\arguments{
\item{risk_contributions}{The risk contributions.}

\item{low_number}{The lowest number of subgroups.}

\item{high_number}{The highest number of subgroups.}

\item{ipw}{a vector of weights per observation to allow for inverse probability of censoring weighting to correct for selection bias}

\item{restore_par_options}{Restore par options.}
}
\value{
A plot of the mean distance by the number of subgroups. The mean distance converges when the optimal number of subgroups are found.
}
\description{
Calculates the mean distance by several number of subgroups to determine the optimal number of subgroups.
}
\examples{
#See the example under CoOL_0_working_example
}
