% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoOL_functions.R
\name{CoOL_6_sub_groups}
\alias{CoOL_6_sub_groups}
\title{Assign sub-groups}
\usage{
CoOL_6_sub_groups(risk_contributions, number_of_subgroups = 3, ipw = 1)
}
\arguments{
\item{risk_contributions}{The risk contributions.}

\item{number_of_subgroups}{The number of sub-groups chosen (Visual inspection is necessary).}

\item{ipw}{a vector of weights per observation to allow for inverse probability of censoring weighting to correct for selection bias}
}
\value{
A vector [number of individuals] with an assigned sub-group.
}
\description{
Calculates presents a dendrogram coloured by the pre-defined number of sub-groups and provides the vector with sub-groups.
}
\examples{
#See the example under CoOL_0_working_example
}
\references{
Rieckmann, Dworzynski, Arras, Lapuschkin, Samek, Arah, Rod, Ekstrom. 2022. Causes of outcome learning: A causal inference-inspired machine learning approach to disentangling common combinations of potential causes of a health outcome. International Journal of Epidemiology <https://doi.org/10.1093/ije/dyac078>
}
