% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoOL_functions.R
\name{CoOL_6_sum_of_individual_effects}
\alias{CoOL_6_sum_of_individual_effects}
\title{Predict the risk based on the sum of individual effects}
\usage{
CoOL_6_sum_of_individual_effects(X, model)
}
\arguments{
\item{X}{The exposure data.}

\item{model}{The fitted the non-negative neural network.}
}
\value{
A value the sum of indivisual effects, had there been no interactions between exposures.
}
\description{
By summing the through the risk as if each individual had been exposed to only one exposure, with the value the individual actually had.
}
\examples{
#See the example under CoOL_0_working_example
}
\references{
Rieckmann, Dworzynski, Arras, Lapuschkin, Samek, Arah, Rod, Ekstrom. 2022. Causes of outcome learning: A causal inference-inspired machine learning approach to disentangling common combinations of potential causes of a health outcome. International Journal of Epidemiology <https://doi.org/10.1093/ije/dyac078>
}
