% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoOL_functions.R
\name{CoOL_7_prevalence_and_mean_risk_plot}
\alias{CoOL_7_prevalence_and_mean_risk_plot}
\title{Prevalence and mean risk plot}
\usage{
CoOL_7_prevalence_and_mean_risk_plot(
  risk_contributions,
  sub_groups,
  title = "Prevalence and mean risk\\nof sub-groups",
  y_max = NA,
  restore_par_options = TRUE,
  colours = NA,
  ipw = 1
)
}
\arguments{
\item{risk_contributions}{The risk contributions.}

\item{sub_groups}{The vector with the sub-groups.}

\item{title}{The title of the plot.}

\item{y_max}{Fix the axis of the risk of the outcome.}

\item{restore_par_options}{Restore par options.}

\item{colours}{Colours indicating each sub-group.}

\item{ipw}{a vector of weights per observation to allow for inverse probability of censoring weighting to correct for selection bias}
}
\value{
A plot with prevalence and mean risks by sub-groups.
}
\description{
This plot shows the prevalence and mean risk for each sub-group. Its distribution hits at sub-groups with great public health potential.
}
\examples{
#See the example under CoOL_0_working_example
}
\references{
Rieckmann, Dworzynski, Arras, Lapuschkin, Samek, Arah, Rod, Ekstrom. 2022. Causes of outcome learning: A causal inference-inspired machine learning approach to disentangling common combinations of potential causes of a health outcome. International Journal of Epidemiology <https://doi.org/10.1093/ije/dyac078>
}
