% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoOL_functions.R
\name{CoOL_9_visualised_mean_risk_contributions_legend}
\alias{CoOL_9_visualised_mean_risk_contributions_legend}
\title{Legend to the visualisation of the mean risk contributions by sub-groups}
\usage{
CoOL_9_visualised_mean_risk_contributions_legend(
  results,
  restore_par_options = TRUE
)
}
\arguments{
\item{results}{CoOL_8_mean_risk_contributions_by_sub_group.}

\item{restore_par_options}{Restore par options.}
}
\description{
Legend to the visualisation of the mean risk contributions by sub-groups. The function uses the output
}
\examples{
#See the example under CoOL_0_working_example
}
\references{
Rieckmann, Dworzynski, Arras, Lapuschkin, Samek, Arah, Rod, Ekstrom. 2022. Causes of outcome learning: A causal inference-inspired machine learning approach to disentangling common combinations of potential causes of a health outcome. International Journal of Epidemiology <https://doi.org/10.1093/ije/dyac078>
}
