% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoOL_functions.R
\name{random}
\alias{random}
\title{Function used as part of other functions}
\usage{
random(r, c)
}
\arguments{
\item{r}{rows in matrix}

\item{c}{columns in matrix}
}
\description{
Function used as part of other functions
}
