% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpprelu_neg}
\alias{rcpprelu_neg}
\title{Function used as part of other functions}
\usage{
rcpprelu_neg(x)
}
\arguments{
\item{x}{input in the negative relu-function}
}
\description{
negative relu-function
}
