% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoOL_functions.R
\name{relu}
\alias{relu}
\title{Function used as part of other functions}
\usage{
relu(input)
}
\arguments{
\item{input}{input in the relu function}
}
\description{
Function used as part of other functions
}
