% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRCL_functions.R
\name{SRCL_0_synthetic_data}
\alias{SRCL_0_synthetic_data}
\title{SRCL synthetic data}
\usage{
SRCL_0_synthetic_data(n)
}
\arguments{
\item{n}{number of observations for the synthetic data}
}
\description{
To reproduce the synthetic data from the paper Synergistic Cause Learning.
}
\examples{

}
