% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRCL_functions.R
\name{SRCL_0_working_example}
\alias{SRCL_0_working_example}
\title{CoOL working example with sex, drug A, and drug B}
\usage{
SRCL_0_working_example(n)
}
\arguments{
\item{n}{number of observations for the synthetic data}
}
\description{
To reproduce the CoOL working example with sex, drug A, and drug B.
}
