% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRCL_functions.R
\name{SRCL_2_train_neural_network_with_confounder}
\alias{SRCL_2_train_neural_network_with_confounder}
\title{Training the monotonistic neural network with a confounder connected to the output layer}
\usage{
SRCL_2_train_neural_network_with_confounder(
  X,
  Y,
  C,
  model,
  lr = 0.01,
  epochs = 50000,
  patience = 500,
  plot_and_evaluation_frequency = 50
)
}
\arguments{
\item{X}{The exposure data}

\item{Y}{The outcome data}

\item{C}{The confounder data}

\item{model}{The fitted monotonistic neural network}

\item{lr}{Learning rate}

\item{epochs}{Epochs}

\item{patience}{The number of epochs allowed without an improvement in performance.}

\item{plot_and_evaluation_frequency}{The interval for plotting the performance and checking the patience}
}
\description{
This function trains the monotonistic neural network with a confounder connected to the output layer. This functions allows one to divide the training process into several steps.
}
\examples{
#See the example under SRCL_0_synthetic_data
}
