% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRCL_functions.R
\name{SRCL_3_plot_neural_network}
\alias{SRCL_3_plot_neural_network}
\title{Plotting the monotonistic neural network}
\usage{
SRCL_3_plot_neural_network(model, names, arrow_size = 2)
}
\arguments{
\item{model}{The fitted monotonistic neural network}

\item{names}{Labels of each exposure}

\item{arrow_size}{defines the arrow_size for the model illustration in the reported training progress.}
}
\description{
This function plots the monotonistic neural network
}
\examples{
#See the example under SRCL_0_synthetic_data
}
