% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRCL_functions.R
\name{SRCL_4_predict_risks}
\alias{SRCL_4_predict_risks}
\title{Predict the risk of the outcome using the fitted monotonistic neural network}
\usage{
SRCL_4_predict_risks(X, model)
}
\arguments{
\item{X}{The exposure data}

\item{model}{The fitted the monotonistic neural network}
}
\description{
Predict the risk of the outcome using the fitted monotonistic neural network.
}
\examples{
#See the example under SRCL_0_synthetic_data
}
