% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRCL_functions.R
\name{SRCL_5_layerwise_relevance_propagation}
\alias{SRCL_5_layerwise_relevance_propagation}
\title{Layer-wise relevance propagation of the fitted monotonistic neural network}
\usage{
SRCL_5_layerwise_relevance_propagation(X, model)
}
\arguments{
\item{X}{The exposure data}

\item{model}{The fitted the monotonistic neural network}
}
\description{
Calculates risk contributions for each exposure and a baseline using layer-wise relevance propagation of the fitted monotonistic neural network and data.
}
\examples{
#See the example under SRCL_0_synthetic_data
}
