% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRCL_functions.R
\name{SRCL_6_individual_effects_matrix}
\alias{SRCL_6_individual_effects_matrix}
\title{Risk contribution matrix based on individual effects (had all other exposures been set to zero)}
\usage{
SRCL_6_individual_effects_matrix(X, model)
}
\arguments{
\item{X}{The exposure data}

\item{model}{The fitted the monotonistic neural network}
}
\description{
Estimating the risk contribution for each exposure if each individual had been exposed to only one exposure, with the value the individual actually had.
}
\examples{
#See the example under SRCL_0_synthetic_data
}
