% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SRCL_cpp_train_network_relu_with_confounder}
\alias{SRCL_cpp_train_network_relu_with_confounder}
\title{Function used as part of other functions}
\usage{
SRCL_cpp_train_network_relu_with_confounder(
  x,
  y,
  c,
  testx,
  testy,
  testc,
  W1_input,
  B1_input,
  W2_input,
  B2_input,
  C2_input,
  lr = 0.01,
  maxepochs = 100
)
}
\arguments{
\item{x}{A matrix of predictors for the training dataset}

\item{y}{A vector of output values for the training data with a length similar to the number of rows of x}

\item{c}{A matrix of predictors for the training data to be regarded as potential confounder(s)}

\item{testx}{A matrix of predictors for the test dataset}

\item{testy}{A vector of output values for the test data with a length similar to the number of rows of x}

\item{testc}{A matrix of predictors for the test data to be regarded as potential confounder(s)}

\item{W1_input}{Input-hidden layer weights}

\item{B1_input}{Biases for the hidden layer}

\item{W2_input}{Hidden-output layer weights}

\item{B2_input}{Bias for the output layer (the baseline risk)}

\item{C2_input}{Weight for the confounder}

\item{lr}{Initial learning rate}

\item{maxepochs}{The maximum number of epochs}
}
\value{
A list of class "SCL" giving the estimated matrices and performance indicators
}
\description{
Used as part of other functions.
}
\author{
Andreas Rieckmann, Piotr Dworzynski, Claus Ekstrøm
}
