% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoOL_functions.R
\name{CoOL_0_binary_encode_exposure_data}
\alias{CoOL_0_binary_encode_exposure_data}
\title{Binary encode exposure data}
\usage{
CoOL_0_binary_encode_exposure_data(exposure_data)
}
\arguments{
\item{exposure_data}{The exposure data set}
}
\description{
This function binary encodes the exposure data set so that each category is coded 0 and 1 (e.g. the variable sex will be two variables men (1/0) and women (1/)).
}
\examples{
#See the example under CoOL_0_synthetic_data

}
