% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoOL_functions.R
\name{CoOL_0_complex_simulation}
\alias{CoOL_0_complex_simulation}
\title{Complex example}
\usage{
CoOL_0_complex_simulation(n)
}
\arguments{
\item{n}{number of observations for the synthetic data.}
}
\value{
A data frame with the columns Y, Physically_active, Low_SES, Mutation_X, LDL, Night_shifts, Air_pollution and n rows.
}
\description{
To reproduce the complex example.
}
\references{
Rieckmann, Dworzynski, Arras, Lapuschkin, Samek, Arah, Rod, Ekstrom. 2022. Causes of outcome learning: A causal inference-inspired machine learning approach to disentangling common combinations of potential causes of a health outcome. International Journal of Epidemiology <https://doi.org/10.1093/ije/dyac078>
}
