% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoOL_functions.R
\name{CoOL_0_confounding_simulation}
\alias{CoOL_0_confounding_simulation}
\title{Confounding example}
\usage{
CoOL_0_confounding_simulation(n)
}
\arguments{
\item{n}{number of observations for the synthetic data.}
}
\value{
A data frame with the columns Y, A, B, C, D, E, F and n rows.
}
\description{
To reproduce the confounding example.
}
\references{
Rieckmann, Dworzynski, Arras, Lapuschkin, Samek, Arah, Rod, Ekstrom. Causes of outcome learning: A causal inference-inspired machine learning approach to disentangling common combinations of potential causes of a health outcome. medRxiv (2020) <doi:10.1101/2020.12.10.20225243>
}
