% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoOL_functions.R
\name{CoOL_0_selection_simulation}
\alias{CoOL_0_selection_simulation}
\title{M-bias example}
\usage{
CoOL_0_selection_simulation(n)
}
\arguments{
\item{n}{number of observations for the synthetic data}
}
\description{
To reproduce the M-bias example.
}
\references{
Rieckmann, Dworzynski, Arras, Lapuschkin, Samek, Arah, Rod, Ekstrom. Causes of outcome learning: A causal inference-inspired machine learning approach to disentangling common combinations of potential causes of a health outcome. medRxiv (2020) <doi:10.1101/2020.12.10.20225243>
}
