% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoOL_functions.R
\name{CoOL_0_working_example}
\alias{CoOL_0_working_example}
\title{CoOL working example with sex, drug A, and drug B}
\usage{
CoOL_0_working_example(n)
}
\arguments{
\item{n}{number of observations for the synthetic data}
}
\description{
To reproduce the CoOL working example with sex, drug A, and drug B.
}
\examples{
#See the example under CoOL_0_synthetic_data

}
