% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoOL_functions.R
\name{CoOL_3_plot_neural_network}
\alias{CoOL_3_plot_neural_network}
\title{Plotting the monotonistic neural network}
\usage{
CoOL_3_plot_neural_network(
  model,
  names,
  arrow_size = NA,
  title = "Model connection weights and intercepts",
  restore_par_options = TRUE
)
}
\arguments{
\item{model}{The fitted monotonistic neural network.}

\item{names}{Labels of each exposure.}

\item{arrow_size}{Define the arrow_size for the model illustration in the reported training progress.}

\item{title}{Title on the plot.}

\item{restore_par_options}{Restore par options.}
}
\description{
This function plots the monotonistic neural network
}
\examples{
#See the example under CoOL_0_working_example
}
\references{
Rieckmann, Dworzynski, Arras, Lapuschkin, Samek, Arah, Rod, Ekstrom. Causes of outcome learning: A causal inference-inspired machine learning approach to disentangling common combinations of potential causes of a health outcome. medRxiv (2020) <doi:10.1101/2020.12.10.20225243>
}
