% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoOL_functions.R
\name{CoOL_3_plot_neural_network}
\alias{CoOL_3_plot_neural_network}
\title{Plotting the monotonistic neural network}
\usage{
CoOL_3_plot_neural_network(model, names, arrow_size = 2, title = "Model")
}
\arguments{
\item{model}{The fitted monotonistic neural network}

\item{names}{Labels of each exposure}

\item{arrow_size}{defines the arrow_size for the model illustration in the reported training progress.}

\item{title}{Title on the plot}
}
\description{
This function plots the monotonistic neural network
}
\examples{
#See the example under CoOL_0_synthetic_data
}
