% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoOL_functions.R
\name{CoOL_4_AUC}
\alias{CoOL_4_AUC}
\title{Predict the risk of the outcome using the fitted monotonistic neural network}
\usage{
CoOL_4_AUC(outcome_data, exposure_data, model, title = "Accuracy")
}
\arguments{
\item{model}{The fitted the monotonistic neural network}

\item{X}{The exposure data}
}
\description{
Predict the risk of the outcome using the fitted monotonistic neural network.
}
\examples{
#See the example under CoOL_0_synthetic_data
}
