% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoOL_functions.R
\name{CoOL_4_AUC}
\alias{CoOL_4_AUC}
\title{Plot the ROC AUC}
\usage{
CoOL_4_AUC(
  outcome_data,
  exposure_data,
  model,
  title = "Receiver operating\\ncharacteristic curve"
)
}
\arguments{
\item{outcome_data}{The outcome data}

\item{exposure_data}{The exposure data}

\item{model}{The fitted the monotonistic neural network}

\item{title}{Title on the plot}
}
\description{
Plot hte ROC AUC
}
\examples{
#See the example under CoOL_0_working_example
}
