% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoOL_functions.R
\name{CoOL_4_predict_risks}
\alias{CoOL_4_predict_risks}
\title{Predict the risk of the outcome using the fitted non-negative neural network}
\usage{
CoOL_4_predict_risks(X, model)
}
\arguments{
\item{X}{The exposure data.}

\item{model}{The fitted the non-negative neural network.}
}
\value{
A vector with the predicted risk of the outcome for each individual.
}
\description{
Predict the risk of the outcome using the fitted non-negative neural network.
}
\examples{
#See the example under CoOL_0_working_example
}
\references{
Rieckmann, Dworzynski, Arras, Lapuschkin, Samek, Arah, Rod, Ekstrom. 2022. Causes of outcome learning: A causal inference-inspired machine learning approach to disentangling common combinations of potential causes of a health outcome. International Journal of Epidemiology <https://doi.org/10.1093/ije/dyac078>
}
