% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoOL_functions.R
\name{CoOL_5_layerwise_relevance_propagation}
\alias{CoOL_5_layerwise_relevance_propagation}
\title{Layer-wise relevance propagation of the fitted monotonistic neural network}
\usage{
CoOL_5_layerwise_relevance_propagation(X, model)
}
\arguments{
\item{X}{The exposure data}

\item{model}{The fitted the monotonistic neural network}
}
\description{
Calculates risk contributions for each exposure and a baseline using layer-wise relevance propagation of the fitted monotonistic neural network and data.
}
\examples{
#See the example under CoOL_0_synthetic_data
}
