% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoOL_functions.R
\name{CoOL_6_individual_effects_matrix}
\alias{CoOL_6_individual_effects_matrix}
\title{Risk contribution matrix based on individual effects (had all other exposures been set to zero)}
\usage{
CoOL_6_individual_effects_matrix(X, model)
}
\arguments{
\item{X}{The exposure data.}

\item{model}{The fitted the non-negative neural network.}
}
\value{
A matrix [Number of individuals, exposures] with the estimated individual effects by each exposure had all other values been set to zero.
}
\description{
Estimating the risk contribution for each exposure if each individual had been exposed to only one exposure, with the value the individual actually had.
}
\examples{
#See the example under CoOL_0_working_example
}
\references{
Rieckmann, Dworzynski, Arras, Lapuschkin, Samek, Arah, Rod, Ekstrom. 2022. Causes of outcome learning: A causal inference-inspired machine learning approach to disentangling common combinations of potential causes of a health outcome. International Journal of Epidemiology <https://doi.org/10.1093/ije/dyac078>
}
