% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoOL_functions.R
\name{CoOL_7_prevalence_and_mean_risk_plot}
\alias{CoOL_7_prevalence_and_mean_risk_plot}
\title{Prevalence and mean risk plot}
\usage{
CoOL_7_prevalence_and_mean_risk_plot(
  risk_contributions,
  sub_groups,
  title = "Prevalence and mean risk\\nof sub-groups",
  ymax = NA
)
}
\arguments{
\item{risk_contributions}{The risk contributions}

\item{sub_groups}{The vector with the sub-groups}

\item{title}{The title of the plot}
}
\description{
This plot shows the prevalence and mean risk for each sub-group. Its destribution hits at sub-groups with great public health potential.
}
\examples{
#See the example under CoOL_0_working_example
}
