% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoOL_functions.R
\name{CoOL_8_mean_risk_contributions_by_sub_group}
\alias{CoOL_8_mean_risk_contributions_by_sub_group}
\title{Mean risk contributions by sub-groups}
\usage{
CoOL_8_mean_risk_contributions_by_sub_group(risk_contributions, sub_groups)
}
\arguments{
\item{risk_contributions}{The risk contributions}

\item{sub_groups}{The vector with the sub-groups}
}
\description{
Table with the mean risk contributions by sub-groups.
}
\examples{
#See the example under CoOL_0_synthetic_data
}
